requirement.txt
asgiref==3.2.10
astroid==2.4.2
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
coverage==5.1
defusedxml==0.6.0
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pylint==2.5.3
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-pythonpath==0.7.3
python3-openid==3.1.0
pytz==2020.1
requests==2.24.0
rope==0.17.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wcwidth==0.2.5
wrapt==1.12.1
